+++ internal/hwmon/gpu.go
@@ -150,6 +150,16 @@ func GpuinfoNvidia(gpudata *data.GPUdata) {
 		}
 		k := strings.TrimSpace(chunks[0])
 		v := strings.TrimSpace(chunks[1])
+
+		if strings.Contains(k, "NVIDIA-SMI has failed") {
+			// an Nvidia card too old for the installed driver
+			cmd := "/bin/env lspci -mm | grep VGA"
+			vgabytes, err := exec.Command("/bin/env", "bash", "-c", cmd).Output()
+			_, gpudata.Name, _ = string.Cut(string(vgabytes), "[")
+			gpudata.Name, _, _ = string.Cut(gpudata.Name, "]")
+			break
+		}
+
 		switch k {
 		case "Product Name":
 			gpudata.Name = strings.TrimPrefix(v, "NVIDIA ")